backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3
python-chess==1.999
requests==2.31.0
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
authlib==1.3.0
httpx==0.26.0
anthropic==0.34.0
tenacity==8.2.3
